- Make testing for SSACSolvers only occur once, to reduce time for subsequent simulations.
