+++ gillespy2/solvers/utilities/solverutils.py
@@ -282,21 +282,3 @@ def dependency_grapher(model, reactions):
 
     return dependent_rxns
 
-"""
-Below is a utility that checks whether or not users can use C++ solvers.
-"""
-
-def check_cpp_support():
-    from gillespy2.solvers.cpp.example_models import Example
-    from gillespy2 import SSACSolver
-    try:
-        model = Example()
-        results = model.run(solver=SSACSolver, cpp_support=True)
-        return True
-    except Exception as e:
-        log.warn('Unable to use C++ optimized SSA: {0}.  The performance of ' \
-        'this package can be significantly increased if you install/configure GCC on ' \
-        'this machine.'.format(e))
-        return False
-
-
